requirements.txt
numpy 
pandas 
scikit-learn 
dvc[s3]
mlflow 
dagshub
python-dotenv